+++ src/main/java/com/dnd/jjakkak/domain/example/repository/ExampleRepository.java
@@ -1,0 +1,15 @@
+package com.dnd.jjakkak.domain.example.repository;
+
+import com.dnd.jjakkak.domain.example.entity.Example;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ * Example Repository Class.
+ *
+ * @author 정승조
+ * @version 2024. 07. 17.
+ */
+public interface ExampleRepository extends JpaRepository<Example, Long> {
+
+
+}
