+++ src/main/java/com/dnd/jjakkak/domain/member/entity/Role.java
@@ -1,0 +1,13 @@
+package com.dnd.jjakkak.domain.member.entity;
+
+/**
+ * 역할입니다.
+ *
+ * @author 류태웅
+ * @version 2024. 07. 20.
+ */
+
+public enum Role {
+    ROLE_USER,
+    ROLE_ADMIN
+}
