--- src/main/java/com/dnd/zzaekkac/domain/example/entity/Example.java
@@ -1,41 +1,0 @@
-package com.dnd.zzaekkac.domain.example.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * Example Entity.
- *
- * @author 정승조
- * @version 2024. 07. 17.
- */
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Example {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String content;
-
-    @Builder
-    public Example(String name, String content) {
-        this.name = name;
-        this.content = content;
-    }
-
-    public void update(String name, String content) {
-        this.name = name;
-        this.content = content;
-    }
-}
