--- src/main/java/com/dnd/zzaekkac/domain/member/jwt/provider/JwtProvider.java
@@ -1,92 +1,0 @@
-package com.dnd.zzaekkac.domain.member.jwt.provider;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
-import java.util.Date;
-
-/**
- * JWT를 만들고 검증하는 Provider입니다.
- *
- * @author 류태웅
- * @version 2024. 07. 24.
- */
-
-@Component
-public class JwtProvider {
-    @Value("${secret.key}")
-    private String secretKey;
-
-    /**
-     * JWT access Token을 만드는 메소드
-     *
-     * <li>특이 사항으로, JWT를 member의 기본키로 생성하는 것이 아님.</li>
-     * <li>kakaoId로 생성함</li>
-     * <li>만료시간 30분</li>
-     *
-     * @param kakaoId String
-     * @return JWT
-     */
-
-    public String createAccessToken(String kakaoId) {
-        Date expiredDate = Date.from(Instant.now().plus(30, ChronoUnit.MINUTES));
-        Key key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-        return Jwts.builder()
-                .signWith(key, SignatureAlgorithm.HS256)
-                .setSubject(kakaoId)
-                .setIssuedAt(new Date())
-                .setExpiration(expiredDate)
-                .compact();
-    }
-
-    /**
-     * JWT refresh Token을 만드는 메소드
-     *
-     * <li>로그아웃에 사용</li>
-     * <li>만료시간 1주일</li>
-     *
-     * @return JWT
-     */
-
-    public String createRefreshToken() {
-        Date expiredDate = Date.from(Instant.now().plus(7, ChronoUnit.DAYS));
-        Key key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-        return Jwts.builder()
-                .signWith(key, SignatureAlgorithm.HS256)
-                .setIssuedAt(new Date())
-                .setExpiration(expiredDate)
-                .compact();
-    }
-
-    /**
-     * JWT를 검증하는 메소드
-     *
-     * @param jwt String
-     * @return subject
-     */
-
-    public String validate(String jwt) {
-        String subject = null;
-        Key key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-        try {
-            Claims claims = Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(jwt)
-                    .getBody();
-            subject = claims.getSubject();
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-        return subject;
-    }
-}
