+++ src/test/java/com/dnd/jjakkak/JjakkakApplicationTests.java
@@ -1,0 +1,13 @@
+package com.dnd.jjakkak;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class JjakkakApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
