--- src/test/java/com/dnd/zzaekkac/ZzaekkacApplicationTests.java
@@ -1,13 +1,0 @@
-package com.dnd.zzaekkac;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ZzaekkacApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
